Ignore certain files only in the top-level
We don't want to ignore other directories called 'dist' that could
exist e.g. within the vendor directory -- that would exclude part of a
dependency accidentally.

A call to 'go build' creates a file called 'openshift-template-tool' in
the top-level that we don't want to see committed, but similarly this
should be limited to the repository root, and not apply recursively.